refactored dialogs out of cards, will need to fix edit button alignment, and passing data back to card from Dialog.
